# backend/services/auth-service/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Wadud Auth Service
  version: 1.0.0
  description: >
    Authentication service for the Wadud telemedicine platform. Issues JWT access
    tokens and rotating refresh tokens, with Redis-backed sessions, OTP, and RBAC.
servers:
  - url: http://localhost:8081
    description: Local dev (direct)
  - url: http://localhost:8080
    description: Local dev (via API gateway)
paths:
  /api/v1/auth/signup:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SignupRequest' }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/SessionEnvelope' } } } }
        '409': { description: Email already registered }
        '422': { description: Validation error }
  /api/v1/auth/login:
    post:
      summary: Authenticate a user
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/SessionEnvelope' } } } }
        '401': { description: Invalid credentials }
  /api/v1/auth/refresh:
    post:
      summary: Rotate refresh token and issue new access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties: { refreshToken: { type: string } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/TokensEnvelope' } } } }
        '401': { description: Invalid/expired/reused refresh token }
  /api/v1/auth/logout:
    post:
      summary: Revoke the session for a refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties: { refreshToken: { type: string } }
      responses:
        '204': { description: No Content }
  /api/v1/auth/forgot-password:
    post:
      summary: Begin password reset (issues OTP)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties: { email: { type: string, format: email } }
      responses:
        '202': { description: Accepted (always, to avoid user enumeration) }
  /api/v1/auth/reset-password:
    post:
      summary: Complete password reset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token: { type: string, description: "format: email:code" }
                newPassword: { type: string, minLength: 8 }
      responses:
        '204': { description: No Content }
        '401': { description: Invalid/expired reset token }
  /api/v1/auth/verify-otp:
    post:
      summary: Verify an OTP code
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VerifyOTPRequest' }
      responses:
        '204': { description: No Content }
        '401': { description: OTP invalid or expired }
  /api/v1/auth/me:
    get:
      summary: Get the authenticated user
      security: [ { bearerAuth: [] } ]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/UserEnvelope' } } } }
        '401': { description: Unauthorized }
  /healthz: { get: { summary: Liveness, responses: { '200': { description: OK } } } }
  /readyz:  { get: { summary: Readiness, responses: { '200': { description: Ready }, '503': { description: Not ready } } } }
  /metrics: { get: { summary: Prometheus metrics, responses: { '200': { description: OK } } } }
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  schemas:
    SignupRequest:
      type: object
      required: [firstName, lastName, email, phone, password]
      properties:
        firstName: { type: string }
        lastName: { type: string }
        email: { type: string, format: email }
        phone: { type: string }
        password: { type: string, minLength: 8 }
        role: { type: string, enum: [patient, doctor, admin] }
        country: { type: string }
        language: { type: string }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    VerifyOTPRequest:
      type: object
      required: [phone, code, type]
      properties:
        phone: { type: string }
        code: { type: string }
        type: { type: string, enum: [registration, login, password_reset] }
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        firstName: { type: string }
        lastName: { type: string }
        email: { type: string }
        phone: { type: string }
        role: { type: string }
        isVerified: { type: boolean }
        isActive: { type: boolean }
        country: { type: string }
        language: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    Tokens:
      type: object
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        expiresAt: { type: string, format: date-time }
    Session:
      type: object
      properties:
        user: { $ref: '#/components/schemas/User' }
        tokens: { $ref: '#/components/schemas/Tokens' }
        role: { type: string }
    SessionEnvelope:
      type: object
      properties:
        data: { $ref: '#/components/schemas/Session' }
        message: { type: string }
        success: { type: boolean }
        timestamp: { type: string, format: date-time }
    TokensEnvelope:
      type: object
      properties:
        data: { $ref: '#/components/schemas/Tokens' }
        message: { type: string }
        success: { type: boolean }
    UserEnvelope:
      type: object
      properties:
        data: { $ref: '#/components/schemas/User' }
        message: { type: string }
        success: { type: boolean }